UserLogin: 12:51pm
Played around for hours trying to connect the login ui with the user ui and got nowhere. Farhan is happy to talk in class re topic. I implemented a quick fix, but also could be a long term solution, instead of turning user into a stack, I turned login into a stack. Accessed the data from there. No changes made to user.files but if we decide to run with the login stack, we'll need to pull user code into that .cpp